+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 ### Completed the test code (200 examples), and make it 100% coverage.
 
+### Feature
+
+1. `type: [String, Integer ..]` will generate an array, which's items
+   would be a oneOf combined schema.
+
 ### Fixed:
 
 1. `desc` will override dry's.
@@ -23,9 +28,11 @@
 
 ### Changed:
 
-1. `instance_eval` => `instance_exec` in dsl.rb.
-2. Guard Clause for `generate_docs` and where schema could be defined.
-3. Change signature of `server` in `api`.
+1. Will NOT do `recognize_is_options_in`.
+2. `instance_eval` => `instance_exec` in dsl.rb.
+3. Guard Clause for `generate_docs` and where schema could be defined.
+4. Change signature of `server` in `api`.
+5. Simplify `recursive`s.
 
 ## [1.4.2 & 1.4.3] - 2017/12/11&13 - [view diff](https://github.com/zhandao/zero-rails_openapi/compare/v1.4.1...v1.4.3)